requirements.txt
accelerate==1.12.0
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.14.0
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
pydeck==0.9.1
pyparsing==3.3.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
smmap==5.0.2
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
